+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/Convert.java
@@ -424,6 +424,15 @@ class Convert {
     return null;
   }
 
+  static Object groundOverlayIdToJson(String groundOverlayId) {
+    if (groundOverlayId == null) {
+      return null;
+    }
+    final Map<String, Object> data = new HashMap<>(1);
+    data.put("groundOverlayId", groundOverlayId);
+    return data;
+  }
+
   static @NonNull Messages.PlatformCameraPosition cameraPositionToPigeon(
       @NonNull CameraPosition position) {
     return new Messages.PlatformCameraPosition.Builder()
@@ -550,6 +559,59 @@ class Convert {
 
   private static String toString(Object o) {
     return (String) o;
+  }
+
+    static String interpretGroundOverlayOptions(Object o, GroundOverlayOptionsSink sink, AssetManager assetManager, float density) {
+    final Map<?, ?> data = toMap(o);
+    final Object consumeTapEvents = data.get("consumeTapEvents");
+    if (consumeTapEvents != null) {
+      sink.setConsumeTapEvents(toBoolean(consumeTapEvents));
+    }
+    final Object transparency = data.get("transparency");
+    if (transparency != null) {
+      sink.setTransparency(toFloat(transparency));
+    }
+    final Object width = data.get("width");
+    final Object height = data.get("height");
+    final Object position = data.get("position");
+    final Object bounds = data.get("bounds");
+    if (height != null) {
+      sink.setPosition(toLatLng(position), toFloat(width), toFloat(height), null);
+    } else {
+      if (width != null) {
+        sink.setPosition(toLatLng(position), toFloat(width), null, null);
+      } else {
+        sink.setPosition(null, null, null, toLatLngBounds(bounds));
+      }
+    }
+    final Object anchor = data.get("anchor");
+    if (anchor != null) {
+      final List<?> anchorData = toList(anchor);
+      sink.setAnchor(toFloat(anchorData.get(0)), toFloat(anchorData.get(1)));
+    }
+
+    final Object bearing = data.get("bearing");
+    if (bearing != null) {
+      sink.setBearing(toFloat(bearing));
+    }
+    final Object visible = data.get("visible");
+    if (visible != null) {
+      sink.setVisible(toBoolean(visible));
+    }
+    final Object zIndex = data.get("zIndex");
+    if (zIndex != null) {
+      sink.setZIndex(toFloat(zIndex));
+    }
+    final Object icon = data.get("icon");
+    if (icon != null) {
+      sink.setIcon(toBitmapDescriptor(icon, assetManager, density));
+    }
+    final String groundOverlayId = (String) data.get("groundOverlayId");
+    if (groundOverlayId == null) {
+      throw new IllegalArgumentException("groundOverlayId was null");
+    } else {
+      return groundOverlayId;
+    }
   }
 
   static void interpretMapConfiguration(