+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapOptionsSink.java
@@ -63,4 +63,6 @@ interface GoogleMapOptionsSink {
   void setInitialTileOverlays(@NonNull List<Messages.PlatformTileOverlay> initialTileOverlays);
 
   void setMapStyle(@Nullable String style);
+
+  void setInitialGroundOverlays(Object initialGroundOverlays);
 }